First pass at windows fix for issue 27.
Broken, disabling buildbot, checking in so I can go to an appointment and work from elsewhere afterwards. More information on future checkin, refer to bug report.

+++ Core/ConnectionSML/src/sml_Connection.h
@@ -228,7 +228,7 @@ public:
 	*
 	* @returns A RemoteConnection instance.
 	*************************************************************/
-	static Connection* CreateRemoteConnection(bool sharedFileSystem, char const* pIPaddress, unsigned short port = kDefaultSMLPort, ErrorCode* pError = NULL) ;
+	static Connection* CreateRemoteConnection(bool sharedFileSystem, char const* pIPaddress, int port = kDefaultSMLPort, ErrorCode* pError = NULL) ;
 
 	/*************************************************************
 	* @brief Create a new connection object wrapping a socket.
